feat: Add WorkoutHeader component for enhanced workout management interface
