+++ apps/frontend/app/routes/_dashboard.fitness.$action.tsx
@@ -1,62 +1,32 @@
 import { useAutoAnimate } from "@formkit/auto-animate/react";
-import {
-	Button,
-	Container,
-	Divider,
-	Group,
-	SimpleGrid,
-	Skeleton,
-	Stack,
-} from "@mantine/core";
-import { notifications } from "@mantine/notifications";
-import {
-	CreateOrUpdateUserWorkoutDocument,
-	CreateOrUpdateUserWorkoutTemplateDocument,
-} from "@ryot/generated/graphql/backend/graphql";
-import { changeCase, isString, parseParameters, sum } from "@ryot/ts-utils";
-import clsx from "clsx";
+import { Button, Container, Group, Skeleton, Stack } from "@mantine/core";
+import { changeCase, isString, parseParameters } from "@ryot/ts-utils";
 import { Howl } from "howler";
 import { produce } from "immer";
 import { RESET } from "jotai/utils";
 import { useState } from "react";
-import { Link, useLoaderData, useNavigate } from "react-router";
+import { Link, useLoaderData } from "react-router";
 import { ClientOnly } from "remix-utils/client-only";
 import { $path } from "safe-routes";
-import { match } from "ts-pattern";
 import { useInterval } from "usehooks-ts";
 import { z } from "zod";
-import { displayWeightWithUnit } from "~/components/fitness";
 import {
 	DEFAULT_SET_TIMEOUT_DELAY,
 	ExerciseDisplay,
-	NameAndOtherInputs,
-	RestTimer,
-	StatDisplay,
-	WorkoutDurationTimer,
+	WorkoutHeader,
 	WorkoutModals,
-	deleteUploadedAsset,
 	getProgressOfExercise,
 	usePerformTasksAfterSetConfirmed,
 	useWorkoutModals,
 } from "~/components/fitness.action";
 import {
 	FitnessAction,
-	FitnessEntity,
-	clientGqlService,
 	dayjsLib,
-	openConfirmationModal,
 	postMessageToServiceWorker,
-	queryClient,
-	queryFactory,
 	sendNotificationToServiceWorker,
 } from "~/lib/common";
+import { useUserPreferences } from "~/lib/hooks";
 import {
-	useApplicationEvents,
-	useUserPreferences,
-	useUserUnitSystem,
-} from "~/lib/hooks";
-import {
-	currentWorkoutToCreateWorkoutInput,
 	useCurrentWorkout,
 	useCurrentWorkoutTimerAtom,
 	useMeasurementsDrawerOpen,
@@ -86,10 +56,7 @@ export const meta = ({ data }: Route.MetaArgs) => {
 export default function Page() {
 	const loaderData = useLoaderData<typeof loader>();
 	const userPreferences = useUserPreferences();
-	const unitSystem = useUserUnitSystem();
-	const events = useApplicationEvents();
 	const [parent] = useAutoAnimate();
-	const navigate = useNavigate();
 	const [isSaveBtnLoading, setIsSaveBtnLoading] = useState(false);
 	const [currentWorkout, setCurrentWorkout] = useCurrentWorkout();
 	const [_, setMeasurementsDrawerOpen] = useMeasurementsDrawerOpen();
@@ -109,8 +76,7 @@ export default function Page() {
 	} = useWorkoutModals();
 	const promptForRestTimer = userPreferences.fitness.logging.promptForRestTimer;
 	const performTasksAfterSetConfirmed = usePerformTasksAfterSetConfirmed();
-	const { advanceOnboardingTourStep, isOnboardingTourInProgress } =
-		useOnboardingTour();
+	const { advanceOnboardingTourStep } = useOnboardingTour();
 
 	const isWorkoutPaused = isString(currentWorkout?.durations.at(-1)?.to);
 	const numberOfExercises = currentWorkout?.exercises.length || 0;
@@ -233,234 +199,35 @@ export default function Page() {
 							<WorkoutModals
 								stopTimer={stopTimer}
 								startTimer={startTimer}
-								assetsModalOpened={assetsModalOpened}
-								setAssetsModalOpened={setAssetsModalOpened}
-								timerDrawerOpened={timerDrawerOpened}
 								openTimerDrawer={openTimerDrawer}
 								closeTimerDrawer={closeTimerDrawer}
+								assetsModalOpened={assetsModalOpened}
+								timerDrawerOpened={timerDrawerOpened}
 								toggleTimerDrawer={toggleTimerDrawer}
+								pauseOrResumeTimer={pauseOrResumeTimer}
+								setAssetsModalOpened={setAssetsModalOpened}
 								isReorderDrawerOpened={isReorderDrawerOpened}
+								setSupersetModalOpened={setSupersetModalOpened}
+								currentWorkoutExercises={currentWorkout.exercises}
 								setIsReorderDrawerOpened={setIsReorderDrawerOpened}
 								supersetWithExerciseIdentifier={supersetWithExerciseIdentifier}
-								setSupersetModalOpened={setSupersetModalOpened}
-								pauseOrResumeTimer={pauseOrResumeTimer}
-								currentWorkoutExercises={currentWorkout.exercises}
 							/>
 							<Stack ref={parent}>
-								<NameAndOtherInputs
-									isCreatingTemplate={loaderData.isCreatingTemplate}
-									openAssetsModal={() => setAssetsModalOpened(null)}
+								<WorkoutHeader
+									stopTimer={stopTimer}
+									loaderData={loaderData}
+									isWorkoutPaused={isWorkoutPaused}
+									isSaveBtnLoading={isSaveBtnLoading}
+									numberOfExercises={numberOfExercises}
+									toggleTimerDrawer={toggleTimerDrawer}
+									openReorderDrawer={openReorderDrawer}
+									setIsSaveBtnLoading={setIsSaveBtnLoading}
+									setAssetsModalOpened={setAssetsModalOpened}
+									shouldDisplayWorkoutTimer={shouldDisplayWorkoutTimer}
+									shouldDisplayFinishButton={shouldDisplayFinishButton}
+									shouldDisplayCancelButton={shouldDisplayCancelButton}
+									shouldDisplayReorderButton={shouldDisplayReorderButton}
 								/>
-								<Group>
-									<WorkoutDurationTimer
-										isWorkoutPaused={isWorkoutPaused}
-										isUpdatingWorkout={loaderData.isUpdatingWorkout}
-										isCreatingTemplate={loaderData.isCreatingTemplate}
-									/>
-									<StatDisplay
-										name="Exercises"
-										value={
-											loaderData.isCreatingTemplate
-												? numberOfExercises.toString()
-												: `${
-														currentWorkout.exercises
-															.map((e) => e.sets.every((s) => s.confirmedAt))
-															.filter(Boolean).length
-													}/${numberOfExercises}`
-										}
-									/>
-									<StatDisplay
-										name="Weight"
-										value={`${displayWeightWithUnit(
-											unitSystem,
-											sum(
-												currentWorkout.exercises
-													.flatMap((e) => e.sets)
-													.flatMap((s) =>
-														loaderData.isCreatingTemplate || s.confirmedAt
-															? Number(s.statistic.reps || 0) *
-																Number(s.statistic.weight || 0)
-															: 0,
-													),
-											).toFixed(),
-										)}`}
-									/>
-									<StatDisplay
-										name="Sets"
-										value={sum(
-											currentWorkout.exercises
-												.flatMap((e) => e.sets)
-												.flatMap((s) =>
-													loaderData.isCreatingTemplate || s.confirmedAt
-														? 1
-														: 0,
-												),
-										).toString()}
-									/>
-								</Group>
-								<Divider />
-								<SimpleGrid
-									cols={
-										Number(shouldDisplayWorkoutTimer) +
-										Number(shouldDisplayReorderButton) +
-										Number(shouldDisplayFinishButton) +
-										Number(shouldDisplayCancelButton)
-									}
-								>
-									{shouldDisplayWorkoutTimer ? (
-										<Button
-											radius="md"
-											color="orange"
-											variant="subtle"
-											size="compact-sm"
-											onClick={toggleTimerDrawer}
-										>
-											<RestTimer />
-										</Button>
-									) : null}
-									{shouldDisplayReorderButton ? (
-										<Button
-											radius="md"
-											color="blue"
-											variant="subtle"
-											size="compact-sm"
-											onClick={() => openReorderDrawer(null)}
-										>
-											Reorder
-										</Button>
-									) : null}
-									{shouldDisplayFinishButton ? (
-										<Button
-											radius="md"
-											color="green"
-											variant="subtle"
-											size="compact-sm"
-											loading={isSaveBtnLoading}
-											disabled={isWorkoutPaused}
-											className={clsx(
-												isOnboardingTourInProgress &&
-													OnboardingTourStepTargets.FinishWorkout,
-											)}
-											onClick={() => {
-												if (!currentWorkout.name) {
-													notifications.show({
-														color: "red",
-														message: `Please give a name to the ${
-															loaderData.isCreatingTemplate
-																? "template"
-																: "workout"
-														}`,
-													});
-													return;
-												}
-												openConfirmationModal(
-													loaderData.isCreatingTemplate
-														? "Only sets that have data will added. Are you sure you want to save this template?"
-														: "Only sets marked as confirmed will be recorded. Are you sure you want to finish this workout?",
-													async () => {
-														setIsSaveBtnLoading(true);
-														if (isOnboardingTourInProgress)
-															advanceOnboardingTourStep();
-
-														await new Promise((r) => setTimeout(r, 1000));
-														const input = currentWorkoutToCreateWorkoutInput(
-															currentWorkout,
-															loaderData.isCreatingTemplate,
-														);
-														for (const exercise of currentWorkout.exercises) {
-															queryClient.removeQueries({
-																queryKey:
-																	queryFactory.fitness.userExerciseDetails(
-																		exercise.exerciseId,
-																	).queryKey,
-															});
-														}
-														stopTimer();
-														try {
-															const [entityId, fitnessEntity] = await match(
-																loaderData.isCreatingTemplate,
-															)
-																.with(true, () =>
-																	clientGqlService
-																		.request(
-																			CreateOrUpdateUserWorkoutTemplateDocument,
-																			input,
-																		)
-																		.then((c) => [
-																			c.createOrUpdateUserWorkoutTemplate,
-																			FitnessEntity.Templates,
-																		]),
-																)
-																.with(false, () =>
-																	clientGqlService
-																		.request(
-																			CreateOrUpdateUserWorkoutDocument,
-																			input,
-																		)
-																		.then((c) => [
-																			c.createOrUpdateUserWorkout,
-																			FitnessEntity.Workouts,
-																		]),
-																)
-																.exhaustive();
-															if (
-																loaderData.action === FitnessAction.LogWorkout
-															)
-																events.createWorkout();
-															setCurrentWorkout(RESET);
-															navigate(
-																$path("/fitness/:entity/:id", {
-																	id: entityId,
-																	entity: fitnessEntity,
-																}),
-															);
-														} catch (e) {
-															notifications.show({
-																color: "red",
-																message: `Error while saving workout: ${JSON.stringify(e)}`,
-															});
-															setIsSaveBtnLoading(false);
-														}
-													},
-												);
-											}}
-										>
-											{loaderData.isCreatingTemplate ||
-											loaderData.isUpdatingWorkout
-												? "Save"
-												: "Finish"}
-										</Button>
-									) : null}
-									{shouldDisplayCancelButton ? (
-										<Button
-											radius="md"
-											color="red"
-											variant="subtle"
-											size="compact-sm"
-											onClick={() => {
-												openConfirmationModal(
-													`Are you sure you want to cancel this ${
-														loaderData.isCreatingTemplate
-															? "template"
-															: "workout"
-													}?`,
-													() => {
-														for (const e of currentWorkout.exercises) {
-															const assets = [...e.images, ...e.videos];
-															for (const asset of assets)
-																deleteUploadedAsset(asset);
-														}
-														navigate($path("/"), { replace: true });
-														setCurrentWorkout(RESET);
-													},
-												);
-											}}
-										>
-											Cancel
-										</Button>
-									) : null}
-								</SimpleGrid>
-								<Divider />
 								{currentWorkout.exercises.map((ex, idx) => (
 									<ExerciseDisplay
 										exerciseIdx={idx}
